⚡️ header footer refactor
