+++ src/app/executive/page.jsx
@@ -7,18 +7,24 @@ import Link from 'next/link';
 import WithAuthorization from '@/components/WithAuthorization';
 import ScscStatusPanel from './ScscStatusPanel';
 import DiscordBotPanel from './DiscordBotPanel';
-import { getApiSecret } from '@/util/getApiSecret';
-import { getBaseUrl } from '@/util/getBaseUrl';
+import {
+  fetchBoards,
+  fetchSigs,
+  fetchPigs,
+  fetchSCSCGlobalStatus,
+  fetchMajors,
+  fetchDiscordBotStatus,
+} from '@/util/fetchAPIData';
 import './page.css';
 
 export default async function AdminPanel() {
   const [boards, sigs, pigs, scscGlobalStatus, majors, discordBotStatus] = await Promise.all([
-    fetchBoards(),
+    fetchBoards([3, 4, 5, 6]),
     fetchSigs(),
     fetchPigs(),
-    fetchScscGlobalStatus(),
+    fetchSCSCGlobalStatus(),
     fetchMajors(),
-    fetchDiscordBot(),
+    fetchDiscordBotStatus(),
   ]);
 
   return (
@@ -77,82 +83,4 @@ export default async function AdminPanel() {
       </div>
     </WithAuthorization>
   );
-}
-
-const targetBoardIds = [3, 4, 5, 6];
-async function fetchBoards() {
-  const boardResults = await Promise.all(
-    targetBoardIds.map(async (id) => {
-      const res = await fetch(`${getBaseUrl()}/api/board/${id}`, {
-        headers: { 'x-api-secret': getApiSecret() },
-        cache: 'no-store',
-      });
-      return res.ok ? await res.json() : null;
-    }),
-  );
-  return boardResults.filter(Boolean);
-}
-
-async function fetchSigs() {
-  const res = await fetch(`${getBaseUrl()}/api/sigs`, {
-    headers: { 'x-api-secret': getApiSecret() },
-    cache: 'no-store',
-  });
-  if (!res.ok) return;
-  const sigsRaw = await res.json();
-  const sigsWithContent = await Promise.all(
-    sigsRaw.map(async (sig) => {
-      const articleRes = await fetch(`${getBaseUrl()}/api/article/${sig.content_id}`, {
-        headers: { 'x-api-secret': getApiSecret() },
-        cache: 'no-store',
-      });
-      const article = articleRes.ok ? await articleRes.json() : { content: '' };
-      return { ...sig, content: article.content };
-    }),
-  );
-  return sigsWithContent;
-}
-
-async function fetchPigs() {
-  const res = await fetch(`${getBaseUrl()}/api/pigs`, {
-    headers: { 'x-api-secret': getApiSecret() },
-    cache: 'no-store',
-  });
-  if (!res.ok) return;
-  const pigsRaw = await res.json();
-  const pigsWithContent = await Promise.all(
-    pigsRaw.map(async (pig) => {
-      const articleRes = await fetch(`${getBaseUrl()}/api/article/${pig.content_id}`, {
-        headers: { 'x-api-secret': getApiSecret() },
-        cache: 'no-store',
-      });
-      const article = articleRes.ok ? await articleRes.json() : { content: '' };
-      return { ...pig, content: article.content };
-    }),
-  );
-  return pigsWithContent;
-}
-
-async function fetchScscGlobalStatus() {
-  const res = await fetch(`${getBaseUrl()}/api/scsc/global/status`, {
-    headers: { 'x-api-secret': getApiSecret() },
-    cache: 'no-store',
-  });
-  return res.ok ? await res.json() : '';
-}
-
-async function fetchMajors() {
-  const res = await fetch(`${getBaseUrl()}/api/majors`, {
-    headers: { 'x-api-secret': getApiSecret() },
-    cache: 'no-store',
-  });
-  if (res.ok) return await res.json();
-}
-
-async function fetchDiscordBot() {
-  const res = await fetch(`${getBaseUrl()}/api/bot/discord/status`, {
-    headers: { 'x-api-secret': getApiSecret() },
-    cache: 'no-store',
-  });
-  if (res.ok) return await res.json();
 }
