+++ src/app/pig/[id]/ClientAuthGate.jsx
@@ -27,7 +27,9 @@ export default function ClientAuthGate({ children }) {
       }
     })();
 
-    return () => { cancelled = true; };
+    return () => {
+      cancelled = true;
+    };
   }, [router]);
 
   return (
