+++ src/app/us/(auth)/login/validator.jsx
@@ -18,8 +18,6 @@ export function phoneNumber(phoneNumber, then) {
   then && then(phoneNumber != undefined && /^01[016789][0-9]{3,4}[0-9]{4}$/.test(phoneNumber));
 }
 
-
-
 export function studentID(studentID, then) {
   const match = studentID.match(/^(\d{4})(\d{5})$/);
   if (!match) return then && then(false);
