+++ src/components/CopyButton.jsx
@@ -2,9 +2,12 @@
 import './CopyButton.css';
 
 export default function CopyButton(props) {
-     return (
-     <button onClick={() => navigator.clipboard.writeText(props.link)} className="invite-link-copy">
-     🔗COPY
-     </button>
-   );
+  return (
+    <button
+      onClick={() => navigator.clipboard.writeText(props.link)}
+      className="invite-link-copy"
+    >
+      🔗COPY
+    </button>
+  );
 }
